requirements.txt
dbt-duckdb
apache-airflow
pyspark
